+++ tools/mbop.cpp
@@ -1,3 +1,4 @@
+#include "mbop_cvt.cpp"
 #include "mbop_delmsg.cpp"
 #include "mbop_emptyfld.cpp"
 #include "mbop_foreach.cpp"