+++ tools/mbop_main.cpp
@@ -628,6 +628,8 @@ int cmd_parser(int argc, char **argv)
 		return cgkreset::main(argc, argv);
 	} else if (strcmp(argv[0], "freeze") == 0) {
 		return set_maint::freeze_main(argc, argv);
+	} else if (strcmp(argv[0], "zaddrxlat") == 0) {
+		return zaddrxlat::main(argc, argv);
 	}
 	return simple_rpc::main(argc, argv);
 }