+++ doc/changelog.rst
@@ -1,3 +1,11 @@
+Gromox 3.3.16 (Development)
+===========================
+
+Enhancements:
+
+* mbop: new "zaddrxlat" command
+
+
 Gromox 3.3 (2025-12-27)
 =======================